requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
python-jose[cryptography]
sqlite-utils
python-multipart